refactor: Move `NoCodeCoverageException` to the Throwable namespace and rename the exception
